eliminated circular dependencies from widget
