+++ packages/messenger-widget/src/components/ConfigureProfile/chain/genome/ConfigureGenomeProfile.tsx
@@ -2,11 +2,11 @@ import { useContext } from 'react';
 import { SubmitOnChainProfile } from '../SubmitOnChainProfile';
 import { submitGenomeNameTransaction, validateGenomeName } from './bl';
 import { ConfigureProfileContext } from '../../context/ConfigureProfileContext';
-import { IChain, NAME_TYPE } from '../../bl';
 import { GlobalContext } from '../../../../utils/context-utils';
 import { useNetwork } from 'wagmi';
 import { AuthContext } from '../../../../context/AuthContext';
 import { ethers } from 'ethers';
+import { IChain, NAME_TYPE } from '../common';
 
 export const ConfigureGenomeProfile = (props: IChain) => {
     const { state, dispatch } = useContext(GlobalContext);
